Rename the .so to not clash with the crate
The primary motivation for this is that rustdoc will otherwise
create docs for a "deltachat" crate both for deltachat as well
as for deltachat-ffi, which is not ideal.

As a side-effect it means it would be possible to install dcc and
dcc-rs next to each other.  But I doubt many people will care much
or for long.  The other effect is that any users, aka the node and
python bindings, need to link with a different lib.  I also think
this isn't the worst idea, it makes it clear which version you're
working with.