Add StrExt.strdup() and remove to_string()
The main part is the adding of the StrExt trait and it's strdup
method.

This mainly removes using a new CString.as_ptr() directly in a
function argument, which is a use-after-free poiniter errror.  It
turns out .strdup() is a lot of these cases.  For the remaining cases
I construct the CString manually as it's easier to see what is
happening and a helper function doesn't do that much.

A few places were Path objects, which can be handled using the
OsStrExt trait's .to_c_string() already in dc_tools, which is custom
made for this and does a much better job and preserving the right
path.

Lastly this removes the incosistent naming of the
to_string()/to_cstring() dc_tools naming.  The remaining to_*()
functions take a *libc::c_char where to_cstring() was the only
exception to this.