Fixed manual account configuration
Replaced the nested `if ok_to_continueX` with returns from a closure.
This allowed to move the check for `shall_stop_ongoing` back into the
progress macro.

Reverted the last progress macro call to a manual call of the callback.
This was required because `shall_stop_ongoing` is not available there.
OTOH, this allows the macro to guard against accidental calls with
progress >999 again.