Copy some docs from the C sources
This copies a bunch of API docs from the C sources, without having
these it is really hard to know what the behaviour of the API should
be like when working on these so I think this is pretty important.

A few are properly converted to the rustdoc format but most only have
the references and some light reformatting done.  It does build
without any warnings with `cargo docs -p deltachat_ffi` however.
Improving the formatting can be done later before we get to actually
publishing this version of the docs.

This also isn't 100% complete.  It's very boring and these are just
the ones I needed for #335, again we can improve coverage as we go.

I'm splitting this off from #335 for two reasons: 1) it's boring and
not relevant and makes the diff look bigger. 2) that PR may not land
at all in that shape, depending on other experiments.