Use new logging in context and log macros
This adopts the new logging functionality in the Context and makes the
existing macros use it.

The main thing to note is that the logger is in an RwLock, which
serialises all threads on writing logs.  Not doing that would need an
immutable logger object, which would turn into doing many more
syscalls for each log write and possibly need a per-thread logfile.