move standards.md here
the document standards.md exist since a long time
to get an overview but also a nice reference to get some links.

however, its placement in the android-repo is a bit unlogical,
probably comes from a time where there was only one implementation :)

as ios and desktop use the same standards and as
all acutal implementations are in the core,
we move it to the core repo.

android can have a link to the new place.