Lost (hidden) attachments while location streaming
While streaming locations, image or speech messages will not shown to the recipient. The messages will be received, the blob is extracted, the location updated, but the message not turning up in the chat of the recipient. Everything looks fine in the senders chat.

Location messages should be hidden from the chat, see #1116 . It's not enough to check the parts of the MIME mail, since it seems to be one anytime. Locations are always sent with type `Text` otherwise the type matches to the attachment like type `Image`.

With the modification location updates are hidden but images will be shown in the chat at the recipient. (Verified with core cli and a deltachat-desktop using the same db file.) Please have a closer look at this, since I'm not very familiar wit the core concepts.