Call update_device_chats automatically during configure
update_device_chats() takes about 2 seconds on a modern device (Android) because the
welcome image file has to be written to the disk as a blob. The problem
was that this was done after the progress bar had vanished and before
anything else happened so that I thought that something had gone wrong
multiple times.

There is no adaption to the UIs necessary because only the first call to
update_device_chats() does anything at all; subsequent calls are
ignored.